requirements.txt
appdirs==1.4.1
awsebcli==3.10.0
blessed==1.14.1
botocore==1.5.17
cement==2.8.2
colorama==0.3.7
Django==1.10.5
django-extensions==1.7.7
django-filter==1.0.1
django-organizations==0.9.1
djangorestframework==3.5.4
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
google-api-python-client==1.6.2
httplib2==0.10.3
jmespath==0.9.1
Markdown==2.6.8
oauth2client==4.0.0
packaging==16.8
pathspec==0.5.0
psycopg2==2.6.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.1.10
python-dateutil==2.6.0
PyYAML==3.12
requests==2.9.1
rsa==3.4.2
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
uritemplate==3.0.0
wcwidth==0.1.7
websocket-client==0.40.0
